Add support for querying tags for files

+++ GraphleManager/src/main/java/com/graphle/graphlemanager/Time/TimeRange.java
@@ -1,4 +1,4 @@
 package com.graphle.graphlemanager.Time;
 
-public record TimeRange(int oldest, int newest) {
+public record TimeRange(long oldest, long newest) {
 }
